move new tests after develop merge
