+++ src/main/webapp/app/admin/lecture-transcription-ingestion/service/lecture-transcription.service.ts
